event-scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.4.0
python-dateutil==2.8.2
celery==5.3.1
redis==4.6.0
